+++ Packs/ServiceNow/Integrations/ServiceNow/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
+Deprecated - We recommend using ServiceNow v2 instead (available from 5.0.0).
 
+## [20.4.1] - 2020-04-29
+-
 
 ## [20.3.3] - 2020-03-18
 -